add build_meta_output and announcement tasks. add logging
The build_meta_output task creates a meta file for the current build and saves it to a public location for the deployment. This allows deployers and developers to see details about their deployment including:

  - Timestamp
  - Release number of deployer php deployment
  - Git revision id
  - Link to git revision
  - Link to PR revision against master branch

Announcements in the command line can be used in verbose and very very verbose mode and include a link to the deployed revision and the launch url of the app if it has been provided in the servers.yml file as `launch_url`.  Slack notifications can also be made via a webhook url and are included as a default task performed after the 'success' task.

Logging has been enabled by providing a default value for `log_file`.