env_symlinks task checks target file is 0 bytes
because a proceeding task uses the touch command on shared files,
which causes a new, empty file to be created if it didn't exist.
this means that if a config file is deleted, the deployer could
create an empty config file in its place which would not produce
an error, which is should, as the results behavior matches that
of a missing config file.