add tasks to fix php52 permission problems
with the addition of docker to our environment, we started getting permissions errors.
our deployment folder is shared between the host server and docker. we deploy code to
the host server and it's server from docker. when the folder is mounted the file
owner/groups are wrong (not set to www-data:www-data). if we ssh into the docker system
and change the owner/groups, the owner changes on the host machine to ubuntu, which
blocks the deployer user's access to correctly run it's deploy tasks.

all new behaviors only affect deployments that contain the string `php52` in the
`host` field of the `servers.yml`. Example: `dep deploy nonprod-php52`

this will:
  - change the owner of the host machine deployer folder to deployer:deployer
  before any other task
  - change to owner of the docker machine deployer folder to deployer:deployer near the
  end of all other tasks

to make this work:
  - the `deployer` user must be in the sudoer group: `sudo usermod -aG sudo deployer`
  - set permissions in `/etc/sudoers` to
  `deployer ALL=(ALL) NOPASSWD: /bin/chown, /usr/bin/docker`
  - deployer user must be added to ubuntu group: `sudo usermod -aG ubuntu deployer`